refactor: blood glucose to body namespace
