+++ unity-project/Assets/Scripts/Body/BloodGlucose.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7e2a826751536e341b02ce18c7e94507
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
