+++ unity-project/Assets/Scripts/Body/UI/BloodGlucose.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bd40574762f76af48a422ec6adc15582
+guid: f845d26c5ad997f4a9a7b685f0d361c2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
