+++ unity-project/Assets/Scripts/Controllers/Indicators.cs
@@ -18,7 +18,7 @@ public class Indicators : MonoBehaviour{
 
     SaveState state;
 
-    public TMP_Text weightText, bmiText, restingHeartRateText, glycatedHemoglobinText, caloriesText;
+    public TMP_Text weightText, bmiText, restingHeartRateText, caloriesText;
 
     [Space]
     public GameObject popupHolder;
@@ -34,7 +34,6 @@ public class Indicators : MonoBehaviour{
         weightText.text = "Peso: " + state.currentWeightKg.ToString("0.00") + "kg";
         bmiText.text = "IMC: " + state.bmi.ToString("0.00");
         restingHeartRateText.text = "FCD: " + state.restingHeartRate;
-        glycatedHemoglobinText.text = "GS: " + state.bgp + "mg/dL";
 
         while(messages.Count > 0){
             PopUpMessage(messages[0]);
